very very change

+++ DataProtection/DataProtection/PageModels/Lab3/ElGamalSignatureModel.cs
@@ -2,24 +2,25 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Net.Sockets;
+using Org.BouncyCastle.Math;
 using System.Threading.Tasks;
 
 namespace DataProtection.PageModels.Lab3
 {
     public class ElGamalSignatureModel
     {
-        public long p { get; set; }
-        public long q { get; set; }
-        public long x { get; set; }
-        public long y { get; set; }
-        public long g { get; set; }
-        public long r { get; set; }
-        public long k { get; set; }
-        public long k_revers { get; set; }
-        public long u { get; set; }
-        public long s { get; set; }
-        public long p_prev { get; set; }
-        public long h { get; set; }
+        public BigInteger p { get; set; }
+        public BigInteger q { get; set; }
+        public BigInteger x { get; set; }
+        public BigInteger y { get; set; }
+        public BigInteger g { get; set; }
+        public BigInteger r { get; set; }
+        public BigInteger k { get; set; }
+        public BigInteger k_revers { get; set; }
+        public BigInteger u { get; set; }
+        public BigInteger s { get; set; }
+        public BigInteger p_prev { get; set; }
+        public BigInteger h { get; set; }
 
     }
 }
